Fix warning due to change in EventMetaData members (unsigned long ... -> int)
git-svn-id: https://belle2.cc.kek.jp/svn/trunk/software@22375 b38c6c8f-bee8-465d-8448-9abe6915d496

+++ alignment/modules/MagnetSwitcher/src/MagnetSwitcherModule.cc
@@ -50,7 +50,7 @@ void MagnetSwitcherModule::beginRun()
   switcher = true;
   // If current run in list, switch field OFF
   for (unsigned int i = 0; i < m_expList.size(); i++) {
-    if (emd->getExperiment() == (unsigned int) m_expList[i] && emd->getRun() == (unsigned int) m_runList[i]) {
+    if (emd->getExperiment() == m_expList[i] && emd->getRun() == m_runList[i]) {
       switcher = false;
       break;
     }
